+++ gluecodium/src/main/java/com/here/gluecodium/generator/dart/DartGenerator.kt
@@ -181,7 +181,7 @@ internal class DartGenerator : Generator {
         injectAsyncHelpers(ffiReferenceMap, asyncHelpers)
 
         val generatedFiles = dartFilteredModel.topElements.flatMap {
-            listOfNotNull(
+            listOf(
                 generateDart(
                     it, dartResolvers, dartNameResolver, listOf(importsCollector, declarationImportsCollector),
                     exportsCollector, typeRepositoriesCollector, predicatesMap, descendantInterfaces,
@@ -216,15 +216,15 @@ internal class DartGenerator : Generator {
         predicates: Map<String, (Any) -> Boolean>,
         descendantInterfaces: Map<String, List<LimeInterface>>,
         asyncHelpers: DartAsyncHelpers.AsyncHelpersGroup?
-    ): GeneratedFile? {
-        val contentTemplateName = selectTemplate(rootElement) ?: return null
+    ): GeneratedFile {
+        val contentTemplateName = selectTemplate(rootElement)
 
         val packagePath = rootElement.path.head.joinToString(separator = "/")
         val fileName = dartNameResolver.resolveFileName(rootElement)
         val filePath = "$packagePath/$fileName"
         val relativePath = "$SRC_DIR_SUFFIX/$filePath.dart"
 
-        val allTypes = LimeTypeHelper.getAllTypes(rootElement).filterNot { it is LimeTypeAlias }
+        val allTypes = LimeTypeHelper.getAllTypes(rootElement)
         val nonExternalTypes = allTypes.filter { it.external?.dart == null }
         val allSymbols = nonExternalTypes.filterNot { CommonGeneratorPredicates.isInternal(it, DART) }
         if (allSymbols.isNotEmpty()) {
@@ -529,7 +529,7 @@ internal class DartGenerator : Generator {
             is LimeEnumeration -> "dart/DartEnumeration"
             is LimeException -> "dart/DartException"
             is LimeLambda -> "dart/DartLambda"
-            is LimeTypeAlias -> null
+            is LimeTypeAlias -> "dart/DartTypeAlias"
             else -> throw GluecodiumExecutionException(
                 "Unsupported top-level element: " +
                     limeElement::class.java.name